+++ whr/__init__.py
@@ -12,7 +12,7 @@ from importlib.metadata import PackageNotFoundError, version
 from whr.game import Game
 from whr.player import Player
 from whr.playerday import PlayerDay
-from whr.utils import UnstableRatingException
+from whr.utils import NoDrawsWarning, UnstableRatingException
 from whr.whole_history_rating import WHR, Base
 
 try:
@@ -24,6 +24,7 @@ __all__ = [
     "WHR",
     "Base",
     "Game",
+    "NoDrawsWarning",
     "Player",
     "PlayerDay",
     "UnstableRatingException",